+++ tests/testthat/test-multistar_as_flat_table.R
@@ -23,7 +23,7 @@ test_that("multistar_as_flat_table works", {
   )
   expect_equal(length(ft[[1]]), 24)
 
-  ms <- dimensional_query(ms_mrs) %>%
+  ms <- dimensional_query(ms_mrs_test) %>%
     select_dimension(name = "where",
                      attributes = c("city", "state")) %>%
     select_dimension(name = "when",
@@ -36,7 +36,6 @@ test_that("multistar_as_flat_table works", {
       name = "mrs_cause",
       measures = c("pneumonia_and_influenza_deaths", "other_deaths")
     ) %>%
-    filter_dimension(name = "where", city == "Boston") %>%
     run_query()
   ft <- ms %>%
     multistar_as_flat_table()
@@ -55,5 +54,5 @@ test_that("multistar_as_flat_table works", {
       "mrs_cause_nrow_agg"
     )
   )
-  expect_equal(length(ft[[1]]), 9)
+  expect_equal(length(ft[[1]]), 6)
 })